Final Project/requirements.txt
django
django-webpack-loader
django-cors-headers
djangorestframework
cx-Oracle
